patches: sed-quote: add new patch
See https://harald.hoyer.xyz/2015/03/05/libtool-getting-rid-of-180000-sed-forks/

Interestingly, at the time, we added the patch to sys-devel/libtool in 5f4225eb727dcd8cc0eebd60620d4b2b160712aa,
but we never pulled it into elt-patches.

Note that we could try to pull in two more fixes:
* 'libtool: optimizing options-parser hooks' (https://git.savannah.gnu.org/cgit/libtool.git/commit/?id=16dbc070d32e6d4601cb5878dfdf69f2e29c84e1)
* 'funclib: refactor quoting methods a bit' (https://git.savannah.gnu.org/cgit/libtool.git/commit/?id=9187e9a231e0a06cc29c336857e95f07f855b2c9)

Unconvinced it's worth it for those though: they're a lot bigger and they affect
the API of the functions so more care is needed with any risk of fuzz, etc.

Bug: https://bugs.gnu.org/20006
Bug: https://bugs.gentoo.org/542252
Signed-off-by: Sam James <sam@gentoo.org>